requirements.txt
gunicorn>=20.1.0
Django>=3.2.7
python-dotenv>=0.19.0
google-api-python-client>=2.21.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6
google-cloud-service-directory>=1.0.2
google-cloud-logging>=2.6.0
requests>=2.26.0
